% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{find_representative_answers}
\alias{find_representative_answers}
\title{Find the top representative columns and answers for each latent class}
\usage{
find_representative_answers(lambda, category_prob, top_n = 10)
}
\arguments{
\item{lambda}{a vector of probabilities for each category.}

\item{category_prob}{a list of a list of a named vector with probabilties
for each answer, latent class and possible category. This
is usually handed over from the result of a call to \code{mixdir()}}

\item{top_n}{the number of top answers per category that will be returned. Default: 10.}
}
\value{
A data frame with four columns: column, category, class and probabilty.
  The probability column contains the chance that an observation belongs to
  the latent class if all that is known about that observation that
  \code{`column`=`category`}
}
\description{
Find the top representative columns and answers for each latent class
}
\examples{
  data("mushroom")
  res <- mixdir(mushroom[1:30, ], beta=1)
  find_representative_answers(res$lambda, res$category_prob, top_n=3)

}
