% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixdir.R
\name{mixdir}
\alias{mixdir}
\title{Run mixture of Dirichlet model}
\usage{
mixdir(X, latent = 3, max_iter = 100, method = c("vi", "em", "mcmc"))
}
\arguments{
\item{X}{A matrix of size (N_ind x N_quest) that contains the categorical responses.}

\item{latent}{The number of latent factors that are used to approximate the model.}

\item{max_iter}{The maximum number of iterations.}

\item{method}{The method that is used to fit the model, can either be variational inference ("vi"),
expectation maximization ("em") or Gibbs Sampling ("mcmc").}
}
\description{
Run mixture of Dirichlet model
}
