% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_class}
\alias{predict_class}
\title{Predict the class of a new observation.}
\usage{
predict_class(X, lambda, category_prob)
}
\arguments{
\item{X}{a named vector of responses for which to predict the latent class.
Values can be missing or it can just be the values for which data is
available.}

\item{lambda}{a vector of probabilities for each category.}

\item{category_prob}{a list of a list of a named vector with probabilities
for each answer, latent class and possible category. This
is usually handed over from the result of a call to \code{mixdir()}}
}
\value{
A vector of \code{length(lambda)} with the probability that \code{X}
  is in that latent class.
}
\description{
Predict the class of a new observation.
}
\details{
Usually the \code{lambda} and \code{category_prob} from a call to
  \code{mixdir()} are used.
}
\examples{
  data("mushroom")
  X <- as.matrix(mushroom)[1:30, ]

  res <- mixdir(X)

  # Predict Class
  predict_class(mushroom[40, ], res$lambda, res$category_prob)
  predict_class(c(`gill-color`="black"), res$lambda, res$category_prob)
}
