% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{find_predictive_features}
\alias{find_predictive_features}
\alias{find_predictive_features.default}
\alias{find_predictive_features.mixdir}
\title{Find the top predictive features and values for each latent class}
\usage{
find_predictive_features(x, ...)

\method{find_predictive_features}{default}(lambda, category_prob, top_n = 10)

\method{find_predictive_features}{mixdir}(mixdir_obj, top_n = 3)
}
\arguments{
\item{lambda}{a vector of probabilities for each category.}

\item{category_prob}{a list of a list of a named vector with probabilties
for each feature, latent class and possible category. This
is usually handed over from the result of a call to \code{mixdir()}}

\item{top_n}{the number of top answers per category that will be returned. Default: 10.}
}
\value{
A data frame with four columns: column, category, class and probabilty.
  The probability column contains the chance that an observation belongs to
  the latent class if all that is known about that observation that
  \code{`column`=`category`}
}
\description{
Find the top predictive features and values for each latent class
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: find predictive features for generic combinations
of a vector lambda and a list category_prob

\item \code{mixdir}: find_predictive_features find predictive features for
an object of class mixdir
}}

\examples{
  data("mushroom")
  res <- mixdir(mushroom[1:30, ], beta=1)
  find_predictive_features(res$lambda, res$category_prob, top_n=3)

}
